Use List[int] for ref_counting on `BasePagedAttentionCache`, Use `_ref_count_lock` when updating ref_counts for `thread-safety`, Move `RefCounts` class back to `trie_attention_cache.py`
