+++ shortfin/python/shortfin_apps/llm/components/kvcache/trie_attention_cache.py
@@ -8,8 +8,27 @@ from .base_attention_cache import (
     BasePagedAttentionCache,
     CacheAllocationFailure,
     PageAllocation,
-    RefCount,
 )
+
+
+@dataclass
+class RefCount:
+    """
+    A reference counter to replace simple int.
+    """
+
+    count: int = 0
+
+    def increment(self) -> int:
+        self.count += 1
+        return self.count
+
+    def decrement(self) -> int:
+        self.count -= 1
+        return self.count
+
+    def is_empty(self) -> bool:
+        return self.count <= 0
 
 
 @dataclass
